+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
+## [unreleased]
+
+### Added
+
+- "registerComponentsOnce" method for the "ComponentFactory" which just registers components if they haven't already been registered
+
 ## [3.3.0]
 
 ### Added